test(oidc): Adds coverage for basic functionality
